# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ETMS API
  version: "1.0.0"
  description: >
    Employee Transportation Management System REST API. Human-readable contract
    in docs/etms/06-api-spec.md; this file covers the endpoints implemented in the
    backend service. All /v1 routes require a Bearer JWT carrying tenant + permissions.
servers:
  - url: https://api.{region}.etms.app/v1
    variables: { region: { default: dev } }
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
security:
  - bearerAuth: []
paths:
...
  /v1/sites:
    get:
      summary: List sites (RLS-scoped)
      responses: { "200": { description: Sites } }
    post:
      summary: Create a site
      responses: { "201": { description: Created }, "409": { description: Duplicate } }
  /v1/sites/{id}:
    patch:
      summary: Update a site
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Updated }, "404": { description: Not found } }
    delete:
      summary: Soft-delete a site
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "204": { description: Deleted } }
  # /v1/{vehicles,employees,vendors,drivers,cost-centers,shifts,routes,rate-cards}
  # expose the same list/create/update/delete shape as /v1/sites.
...
  /v1/notifications/send:
    post:
      summary: Render a DB template → dispatch on a channel → record delivery
      responses: { "200": { description: Sent/queued }, "422": { description: Unknown template } }
...